+++ MatrixKit/Controllers/MXKViewController.m
@@ -18,6 +18,7 @@
 #import "MXKViewController.h"
 
 #import "UIViewController+MatrixKit.h"
+#import "MXSession+MatrixKit.h"
 
 const CGFloat MXKViewControllerMaxExternalKeyboardHeight = 80;
 
@@ -460,11 +461,7 @@ const CGFloat MXKViewControllerMaxExternalKeyboardHeight = 80;
                 else
                 {
                     allHomeserverNotReachable = NO;
-                    
-                    if (mxSession.state == MXSessionStateSyncInProgress || mxSession.state == MXSessionStateInitialised)
-                    {
-                        isActivityInProgress = YES;
-                    }
+                    isActivityInProgress = mxSession.shouldShowActivityIndicator;
                 }
                 
                 index++;
@@ -548,7 +545,7 @@ const CGFloat MXKViewControllerMaxExternalKeyboardHeight = 80;
     BOOL isActivityInProgress = NO;
     for (MXSession *mxSession in mxSessionArray)
     {
-        if (mxSession.state == MXSessionStateSyncInProgress || mxSession.state == MXSessionStateInitialised)
+        if (mxSession.shouldShowActivityIndicator)
         {
             isActivityInProgress = YES;
         }